+++ examples/python/cig_multiplayer_bots.py
@@ -21,11 +21,16 @@ game.add_game_args("-host 1 -deathmatch +timelimit 10.0 "
                    "+sv_forcerespawn 1 +sv_noautoaim 1 +sv_respawnprotect 1 +sv_spawnfarthest 1 +sv_nocrouch 1 "
                    "+viz_respawn_delay 10 +viz_nocheat 1")
 
+# Bots are loaded from file, that by default is bots.cfg located in the same dir as ViZDoom exe
+# Other location of bots configuration can be specified by passing this argument
+game.add_game_args("+viz_bots_path ../../scenarios/perfect_bots.cfg")
+
 # Name your agent and select color
 # colors: 0 - green, 1 - gray, 2 - brown, 3 - red, 4 - light gray, 5 - light brown, 6 - light red, 7 - light blue
 game.add_game_args("+name AI +colorset 0")
 
 game.set_mode(Mode.PLAYER)
+game.set_console_enabled(True)
 
 #game.set_window_visible(False)