+++ src/lib_python/ViZDoomGamePython.cpp
@@ -75,6 +75,10 @@ namespace vizdoom {
             this->pyState->screenBuffer = this->dataToNumpyArray(colorDims, this->colorShape, NPY_UBYTE, this->state->screenBuffer->data());
         else this->pyState->screenBuffer = pyb::none();
 
+        if (this->state->audioBuffer != nullptr)
+            this->pyState->audioBuffer = this->dataToNumpyArray(2, this->audioShape, NPY_SHORT, this->state->audioBuffer->data());
+        else this->pyState->audioBuffer = pyb::none();
+
         if (this->state->depthBuffer != nullptr)
             this->pyState->depthBuffer = this->dataToNumpyArray(2, this->grayShape, NPY_UBYTE, this->state->depthBuffer->data());
         else this->pyState->depthBuffer = pyb::none();
@@ -179,6 +183,12 @@ namespace vizdoom {
         DoomGame::init();
     }
 
+    void DoomGamePython::newEpisode(std::string filePath) {
+
+        ReleaseGIL gil = ReleaseGIL();  // this prevents the deadlock during the start of multiplayer game, if different Doom instances are started from different Python threads
+        DoomGame::newEpisode(filePath);
+    }
+
     void DoomGamePython::advanceAction(unsigned int tics, bool updateState){
         ReleaseGIL gil = ReleaseGIL();
         DoomGame::advanceAction(tics, updateState);
@@ -188,7 +198,6 @@ namespace vizdoom {
         ReleaseGIL gil = ReleaseGIL();
         DoomGame::respawnPlayer();
     }
-
 
     void DoomGamePython::updateBuffersShapes(){
         int channels = this->getScreenChannels();
@@ -211,6 +220,9 @@ namespace vizdoom {
 
         this->grayShape[0] = height;
         this->grayShape[1] = width;
+
+        this->audioShape[0] = this->getAudioSamplesPerTic() * this->getAudioBufferSize();
+        this->audioShape[1] = 2;
     }