+++ src/lib_python/ViZDoomPythonModule.cpp
@@ -370,6 +370,11 @@ PYBIND11_MODULE(vizdoom, vz){
         ENUM_VAL_2_PYT(PLAYER16_FRAGCOUNT)
         .export_values();
 
+    enum_<SamplingRate>(vz, "SamplingRate")
+        ENUM_VAL_2_PYT(SR_11025)
+        ENUM_VAL_2_PYT(SR_22050)
+        ENUM_VAL_2_PYT(SR_44100)
+        .export_values();
 
     /* Structs */
     /*----------------------------------------------------------------------------------------------------------------*/
@@ -427,6 +432,7 @@ PYBIND11_MODULE(vizdoom, vz){
         .def_readonly("game_variables", &GameStatePython::gameVariables)
 
         .def_readonly("screen_buffer", &GameStatePython::screenBuffer)
+        .def_readonly("audio_buffer", &GameStatePython::audioBuffer)
         .def_readonly("depth_buffer", &GameStatePython::depthBuffer)
         .def_readonly("labels_buffer", &GameStatePython::labelsBuffer)
         .def_readonly("automap_buffer", &GameStatePython::automapBuffer)
@@ -536,6 +542,13 @@ PYBIND11_MODULE(vizdoom, vz){
 
         .def("set_console_enabled",&DoomGamePython::setConsoleEnabled)
         .def("set_sound_enabled",&DoomGamePython::setSoundEnabled)
+
+        .def("is_audio_buffer_enabled",&DoomGamePython::isAudioBufferEnabled)
+        .def("set_audio_buffer_enabled",&DoomGamePython::setAudioBufferEnabled)
+        .def("get_audio_sampling_rate",&DoomGamePython::getAudioSamplingRate)
+        .def("set_audio_sampling_rate",&DoomGamePython::setAudioSamplingRate)
+        .def("get_audio_buffer_size",&DoomGamePython::getAudioBufferSize)
+        .def("set_audio_buffer_size",&DoomGamePython::setAudioBufferSize)
 
         .def("set_screen_resolution", &DoomGamePython::setScreenResolution)
         .def("set_screen_format", &DoomGamePython::setScreenFormat)