+++ src/vizdoom/src/s_sound.h
@@ -340,6 +340,7 @@ void S_SetSoundPaused (int state);
 // Updates music & sounds
 //
 void S_UpdateSounds (AActor *listener);
+void S_GetRender (void *targetBuffer, int bufferLen);
 
 void S_RestoreEvictedChannels();