+++ src/vizdoom/src/sound/i_sound.cpp
@@ -73,11 +73,17 @@ extern HINSTANCE g_hInst;
 #include "i_video.h"
 #include "s_sound.h"
 #include "v_text.h"
+#include "viz_main.h"
 #include "gi.h"
 
 #include "doomdef.h"
 
 EXTERN_CVAR (Float, snd_sfxvolume)
+
+// VIZDOOM_CODE
+EXTERN_CVAR (Bool, viz_soft_audio)
+EXTERN_CVAR (Int, viz_samp_freq)
+
 CVAR (Int, snd_samplerate, 0, CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 CVAR (Int, snd_buffersize, 0, CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 CVAR (String, snd_output, "default", CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
@@ -284,7 +290,7 @@ void I_InitSound ()
 			{
 				Printf (TEXTCOLOR_RED"FMod Ex Sound init failed. Trying OpenAL.\n");
 				I_CloseSound();
-				GSnd = new OpenALSoundRenderer;
+				GSnd = new OpenALSoundRenderer(0);
 				snd_backend = "openal";
 			}
 		#endif
@@ -294,7 +300,11 @@ void I_InitSound ()
 		#ifndef NO_OPENAL
 			if (IsOpenALPresent())
 			{
-				GSnd = new OpenALSoundRenderer;
+                if (*viz_soft_audio){
+                    GSnd = new OpenALSoundRenderer(*viz_samp_freq);
+                } else {
+                    GSnd = new OpenALSoundRenderer(0);
+                }
 			}
 		#endif
 		#ifndef NO_FMOD