+++ src/vizdoom/src/sound/oalsound.cpp
@@ -61,20 +61,28 @@
 CVAR (String, snd_aldevice, "Default", CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 CVAR (Bool, snd_efx, true, CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 
+
 #ifdef _WIN32
 static HMODULE hmodOpenAL;
 #define OPENALLIB "openal32.dll"
 #else
 static void* hmodOpenAL;
-#ifdef __APPLE__
-#define OPENALLIB "OpenAL.framework/OpenAL"
-#else
-#define OPENALLIB "libopenal.so"
-#endif
+#define OPENALLIB "libopenal.so.1"
 #define LoadLibrary(x) dlopen((x), RTLD_LAZY)
 #define GetProcAddress(a,b) dlsym((a),(b))
 #define FreeLibrary(x) dlclose((x))
 #endif
+
+#ifdef __APPLE__
+// User's library (like OpenAL Soft installed manually or via Homebrew) has precedence
+// over Apple's OpenAL framework which lacks several important features
+#define OPENALLIB1 "libopenal.1.dylib"
+#define OPENALLIB2 "OpenAL.framework/OpenAL"
+#else // !__APPLE__
+#define OPENALLIB1 NicePath("$PROGDIR/" OPENALLIB)
+#define OPENALLIB2 OPENALLIB
+#endif
+
 
 bool IsOpenALPresent()
 {
@@ -89,10 +97,10 @@ bool IsOpenALPresent()
 		done = true;
 		if (hmodOpenAL == NULL)
 		{
-			hmodOpenAL = LoadLibrary(NicePath("$PROGDIR/" OPENALLIB));
+			hmodOpenAL = LoadLibrary(OPENALLIB1);
 			if (hmodOpenAL == NULL)
 			{
-				hmodOpenAL = LoadLibrary(OPENALLIB);
+				hmodOpenAL = LoadLibrary(OPENALLIB2);
 				if (hmodOpenAL == NULL)
 				{
 					return false;
@@ -681,21 +689,55 @@ ALCdevice *OpenALSoundRenderer::InitDevice()
 	return device;
 }
 
+// VIZDOOM_CODE
+ALCdevice *OpenALSoundRenderer::InitSoftDevice()
+{
+    ALCdevice *device = NULL;
+    if(!alcIsExtensionPresent(NULL, "ALC_SOFT_loopback"))
+    {
+        Printf(TEXTCOLOR_ORANGE"ALC_SOFT_loopback extensions not supported\n");
+    }
+    else if(!alcIsExtensionPresent(NULL, "ALC_EXT_thread_local_context"))
+    {
+        Printf(TEXTCOLOR_ORANGE"ALC_EXT_thread_local_context not supported\n");
+    }
+    else {
+        device = alcLoopbackOpenDeviceSOFT(NULL);
+    }
+    return device;
+}
+
+void OpenALSoundRenderer::GetRenderBuffer(void *targetBuffer, int numSamples)
+{
+    alcRenderSamplesSOFT(Device, targetBuffer, numSamples);
+}
+
 
 template<typename T>
 static void LoadALFunc(const char *name, T *x)
 { *x = reinterpret_cast<T>(alGetProcAddress(name)); }
 
 #define LOAD_FUNC(x)  (LoadALFunc(#x, &x))
-OpenALSoundRenderer::OpenALSoundRenderer()
+OpenALSoundRenderer::OpenALSoundRenderer(unsigned int sampling_freq)
     : Device(NULL), Context(NULL), SFXPaused(0), PrevEnvironment(NULL), EnvSlot(0)
 {
     EnvFilters[0] = EnvFilters[1] = 0;
 
+    if (sampling_freq > 0) {
+        alcRenderSamplesSOFT = (LPALCRENDERSAMPLESSOFT) alcGetProcAddress(NULL, "alcRenderSamplesSOFT");
+        alcSetThreadContext = (PFNALCSETTHREADCONTEXTPROC) alcGetProcAddress(NULL, "alcSetThreadContext");
+        alcLoopbackOpenDeviceSOFT = (LPALCLOOPBACKOPENDEVICESOFT) alcGetProcAddress(NULL, "alcLoopbackOpenDeviceSOFT");
+    }
     Printf("I_InitSound: Initializing OpenAL\n");
 
-	Device = InitDevice();
-	if (Device == NULL) return;
+    // VIZDOOM_CODE
+    if (sampling_freq > 0) {
+        Device = InitSoftDevice();
+    } else {
+        Device = InitDevice();
+    }
+
+    if (Device == NULL) return;
 
     const ALCchar *current = NULL;
     if(alcIsExtensionPresent(Device, "ALC_ENUMERATE_ALL_EXT"))
@@ -725,7 +767,22 @@ OpenALSoundRenderer::OpenALSoundRenderer()
     // Other attribs..?
     attribs.Push(0);
 
-    Context = alcCreateContext(Device, &attribs[0]);
+    if (sampling_freq > 0) {
+        ALCint attrs[] = {
+                /* Standard 16-bit stereo 44.1khz. Can change as desired. */
+                ALC_FORMAT_TYPE_SOFT, ALC_SHORT_SOFT,
+                ALC_FORMAT_CHANNELS_SOFT, ALC_STEREO_SOFT,
+                ALC_FREQUENCY, static_cast<ALCint>(sampling_freq),
+//            ALC_FREQUENCY, 22050,
+                /* end-of-list */
+                0
+        };
+        Context = alcCreateContext(Device, attrs);
+        alcSetThreadContext(Context);
+    } else {
+        Context = alcCreateContext(Device, &attribs[0]);
+    }
+
     if(!Context || alcMakeContextCurrent(Context) == ALC_FALSE)
     {
         Printf(TEXTCOLOR_RED"  Failed to setup context: %s\n", alcGetString(Device, alcGetError(Device)));
@@ -735,6 +792,10 @@ OpenALSoundRenderer::OpenALSoundRenderer()
         alcCloseDevice(Device);
         Device = NULL;
         return;
+    }
+    if (sampling_freq > 0) {
+        ALuint buffer = 0;
+        alGenBuffers(1, &buffer);
     }
     attribs.Clear();